proyecto-parcial/requirements.txt
pandas
pyyaml
requests
pyarrow
pytest
